Added ANTLR parser option.
You will now need an ANTLR runtime to build the compiler.  Get it from:

	http://www.antlr.org/download.html

If you won't be working with the grammar get the "ANTLR tool+runtime binary only jar".
If you will, I recommend getting ANTLRWorks (which includes the runtime).

Current default is to use the old hand-written parser, as the error handling isn't there in the ANTLR version yet (doesn't even stop).  But to use the ANTLR based parser when you run the compiler, include

     -XDparser=vn