Fix JFXC-710 : Compiler support for 'override attribute' to enable defaults on superclass attributes This is the last piece:  add the syntax, hook it up, and add a test. Syntax is:
             override attribute foo = 14;
             override attribute foo on replace {...}  // or full slice syntax for sequences
             override attribute foo = 99 on replace {...}

Note that the initial value only gets set once now, no matter if its origin is an object literal, an explicit default on an attribute, a default default on an attribute, or an override.  And in all these cases the value is set (or bind installed) in the order the attributes appear in the class, superclass to subclass.