JFXC-786: JavaFX Runtime filter package should be built upon Effects Framework
This is just the initial commit to get the ball rolling.  Will
need more testing.  Includes a bunch of changes to make
Node.fx use the FXNode class from Scenario, which greatly
simplifies management of the filter chain.  As an added bonus,
the FXNode class is optimized to collapse itself when
filter properties go unused, which should lead to significant
footprint savings for FX apps (and perhaps a little performance
win as well).