Fixed JFXC-905 : Binding Overhaul: switch bound attributes and object literals to use overhaul Fixed JFXC-866 : Binding Overhaul: switch bound local variables to use overhaul Fixed JFXC-988 : Binding Overhaul: FXPad Fixed JFXC-992 : Binding Overhaul: StudioMoto Fixed JFXC-1008 : Binding Overhaul: verification error
This commit fixes the verification error, which was the last remaining issue for the binding overhaul working with both JavaFXPad and StudioMoto, which, in turn, were the last remaining issues before flicking the switch and enabling the binding overhaul for bound expressions.

The Binding Overhaul is now enabled for bound expressions in local variable initializers, attribute default initializers, and object literal attribute initializers.

Enabling the Binding Overhaul for bound functions is not yet addressed and is covered by other issues.

Fixed JFXC-978 : Binding Overhaul: NPE in back-end attribution
The NPE work-around approach is now fully functional, so, until such time as we are addressing optimizations (and this would be low amongst those) the work-around will be the solution.

The F15 feature test is seeing excess updates, these should be addressed when bound functions are moved to the Binding Overhaul and by other update fixes.  For now, the expected result includes the excess updates.