Fixed JFXC-784 : Binding Overhaul Fixed JFXC-865 : Binding Overhaul: only generate bound function version when 'bound' Fixed JFXC-863 : Binding Overhaul: JavafxMethodSymbol has isBound() Fixed JFXC-501 : Binding Overhaul: Bound sequence updates Fixed JFXC-691 : Problem with function local variables when bound
Correct semantics of function binding now enabled.
Removing the bogus re-execute the whole function semantics is the last piece needed to complete the Binding Overhaul.
Flicking the switch is possible thanks to Brian's fixes to the update semantics in the compiler runtime.

javafx.ui version of JavaFXPad now works with this enabled.
As Chris has reported Reprise demos work with it enabled.
javafx.ui version of StudioMoto comes up, though the bottom panel is missing its components.

Given that Reprise and JavaFXPad work, the weight is on the side of flicking the switch.

Regression test added for 501.
691 now working so moved out of currently-failing.