Fixed JFXC-913 : Accessing bound variable inside a function causes compiler crash
There was a mixed approach that was being used.  For top-level vars moved to the module a beheaded attribute initialization was being attempted with the real initialization in the run method.   The bogus attribute initialization was under some conditions crashing the compiler.

Change them to full attributes with standard attribute initialization.  But this requires moving any other top-level var to the module if it is used in the initialization of a moved var.

Also, remove an unused var and its set-up from Attr.