Workaround for accepted inheritance-cycles (JFXC-1101 and JFXC-1149).
The workaround takes advantage of the fact, that all classes in the cycle are in the same package. This is true for Reprise, but there are probably still cycles possible, which span over two or more packages. To fix this completely, we have to wait for the official fix in javac.