Partial access control re-write.
The class file reader now skip most synthetic fields and methods.
A new Inherited annotation is used to "de-select" generated
stub fields and methods for inherited methods and attributes,
so these can also be skipped by teh class file reader.
"Attribute" access annotations are now on fields, not methods.
The ClassSymbol for a source file and from a class-file are
now much closer.  Because of that, I was able to rid of some of
the separate handling between classes-from-source and
classes-from-classfiles.

Also, some minor optimization (use Name rather than String)
while I was at it.

This doesn't do name mangling of attributes, but the attributeName
method sets that up.

Also still to do is tweaking of the access modifiers as discussed
- i.e. module visibility, plus more testing.