JFXC-1351 JFXC-1537 JFXC-1532 JFXC-1551
There is a issue with Vista with Aero that means we can not change a window from opaque to transparent on the fly. Because of this we have changed the way the transparency works in FX so that now if you want to make a window transparent you have to set the windowStyle attribute to WindowStyle.TRANSPARENT when you create the Window (Frame or Dialog). Changing the transparency of the fill of the Stage no longer has any effect on the transparency of the window.