Begin the process of building v4 of the JavaFX lexer/parser.
This version will remove the need for an ANTLR AST walker and produce
the JavaFX AST directly from the parse. This requires moving some of the
helper methods from AbstractGeneratedTreeParser.java into
AbstractGeneratedParser.java and creating a new ANTLR grammar v4.g
which will eventually take over from v3.g and v3Walker.g.

Note that this commit only adds v4.g to the repository so it is tracked
and does not replace the current v3 parser. The v4.g parser generates
java code using the ANTLR 3.1 tool but the resulting code is not yet
able to compile. You can safely ignore it for now.