Final typos and ruler reference misnomers fixed for v4 grammar. Generated java now compiles cleanly.
Now that we are back to square one, albeit with a grammar that should build ASTs except for catastrophic errors, we can run regression tests and then switch to the v4 grammar at our lesiure. This switch will happen some time this week, and at that point perhaps it is a good time to pull the plug on constructs and keywords that are removed from the final language spec.

Next begins the process of error recovery and reporting improvements to support the downstream tools such as code completion much more cleanly.