Fixed JFXC-1926 : Statically import top-level functions in java.lang Currently these functions are in java.lang.FX -- awaiting a final decision on placement. But it doesn't much matter, they are all statically imported so the class isn't of any consequence.
Fixed JFXC-1924 : println
Added print(Object)  and println(Object).
Because of JFXC-1926 (above) they are ready for use, no import needed:   println("hi")
With regression test.

Partial JFXC-1249 : Provide isInitialized() method for variables
Library function and the needed special processing during translation are in place.
It calls into AbstractVariable in the compiler runtime, but that isn't giving the right value yet, see JFXC-1929.
Also checks need to be added to assure only instance variable references are passed in.
Regression test added, but it is in currently-failing until the compiler runtime issue is fixed.