Improvements for JFXC-1376 : Ensure that when an required identifier or name is missing from the source code, a fake node is produced whose identifer name/text is always:    <missing IDENTIFIER> which is not a valid identifer name and so is easilly identifed when walking the AST.
Consider the inpput:

import javafx.gui.*;

class Box extends Rectangle {

 var g = indexof
 var

}

The compiler now issues the error messages:

t4.fx:5: Sorry, I was trying to understand an identifier but I got confused because you seem to have missed out 'identifier' which is an identifier that should be there.
 var g = indexof
                ^
t4.fx:8: Sorry, I was trying to understand an identifier but I got confused when I saw '}'.
 Perhaps you are missing a 'identifier'.
}
^

But, when you examine the AST, it looks like this:

import javafx.gui.*;

class Box {
    var g = indexof <missing IDENTIFIER>;
    var <missing IDENTIFIER>;

}


Looking for feedback as to whether this is an appropriate way for the IDE tools
to discover this, or whether an error node is preferred (more difficult to engineer
but I will do it if preferred.)