Infrastructure for missing nodes in the AST.  Which support in the API -- query with isMissing().
With two cases: one for a concrete missing node (identifer), and one for an abstract missing node (expression).
Can be extended to other nodes by hum a few bars.

Parser changed to create a JFXMissingIdent when missing

Test to show that missing identifier is indeed marked as missing, and that nothing else is.