Fixed (I'll assume) JFXC-1895 : javafxdoc doesn't recognize public-init or public-read access modifiers I have added the following to FieldDoc(Impl) --
  isPublicRead()
  isPublicInit()
  isDef()

The last, though not requested, you will need to report in the docs too.

In the process I noted that the general access control information isPackagePrivate() was incorrect (note the docs should indicate this as 'package'), I have fixed isPackagePrivate().  Also there was no:

  isScriptPrivate()

So, I have added that to ProgramElementDoc(Impl).

And, the modifiers() method was return a wrong string, so I have fixed this too.

I have added the corresponding methods to XMLDoclet.java since, sadly, there is no interface.

And, finally, I have tested none of this, since I don't have that set-up.  Let me know if anything is broken.