fix for JFXC-1899: javafxdoc doesn't sort the list of packages and fix for JFXC-1914: javafxdoc output seems currently to be biased towards the SDK documentation
I added two xslt parameters.  sdk-overview will turn on the standard overview boilerplate text with the JavaFX platform diagram. Since this is really only appropriate for the SDK itself, this parameter is off by default.  The other new parameter is 'std.doctitle' which sets the name of the docs displayed on the main index page.

I've also added sorting to the package list in the main index page.