Fixed JFXC-1375
Althoguh this issue was fixed by prior checkins, the Attribution code does not
yet cater for the fact that some nodes may be visited as part of beloning to
ERRONEOUS nodes and so incorrectly asserts such things in variaous places. I
am now in the process of finding and detecting such things.