Progress JFXC-1117
Prevent AST walkers from asserting unreachable nodes as they are now
reachable by visiting an Erroneous node that encapsulates them.

Correct the DiagnosticPosition notification for the error regarding
garbled class declaration elements to illustrate that the parser can,
and will, do a much better job of telling the diagnostic listner
about the span of the error and not just the start point of it. I will
dealy improving the rest of the messages until remaining tree walking
aborts are fixed.