Progress JFXC-1117
Update the diagnostic positions of those parser errors generated
by the parser, because they were easy to do and make a huge difference to the
way the IDE plugin looks when errors are found. More to come on that.

Don't allow visitErroneous (javaFxAttr) to try and visit erroneous nodes
as a lot of the code does not support being visited as part of an Erroneous
block, and there is little point automatically visiting thse nodes. The IDE
can decode to attribute these nodes if it feels it can glean anything from them