Fixed JFXC-2020 Fixed JFXC-2015
Provide a generic mechansim to report errors with reference to the AST we are producing,
for those cases where the IDE wants to know what the error point was in terms of the node and not
the token that causes the error to be detected. So in the case of:

var x :
}

The error will be raised at the ':' exactly, even though the error message will correctly
state that the '}' was not what it as looking for. This should greatly aid the IDE plugin
developers.

Also, the parser now reports errors via a DiagnosticPostion and not just a simple offset. These
errors are accumualted in Erroneous nodes taht would not otherwise be produced, and it is
these nodes that the error is logged against. This means that hte diagnostic listner used by
the IDE will get a true input stream span, and as a result, it now succesfully underlines
in red, all the paresr errors, with the exception of where we are missig a token and the span
we give to the error contains only whitespace. The IDE underlining routine shoudl be
modified to cope with that by underlining whitespace if the span is ONLY whitespace, which
should cover all cases.

The error nodes accumulated for syntax errors are given to the AST as the last set of script
elements, which means they apeare as a sequence at the end of the synthesized main function
for the script. This might not be ideal and perhaps we should place them at the COMPILATION_UNIT
level in the same manner that package and imports are done. Waiting for feedback from IDE developers
for that one.

Finally, I implemented the detection of missing intializers for 'def' declartions in the
parser, as otherwise the AST has lost information about location and cannot tell the
IDE just exactly where the initialization should be inserted. With this change, an error and an
Erroneous node are produced with the correct locations, and the IDE could even fill in some
default value now, such as = 0 if defined as an integer, and so on. Error message provided
for this.