Fixed JFXC-2030
A class definition that is found to be erroneous for some fairly
innocuous reason, such as having the closing '}' missing or some
other syntactical problem that otherwise builds a list of defintions
anyway, then the AST for the class definition is now produced and
an erronoeus nodes is entered intot he list of script level provided
erroneous nodes. The erroneous node is also added to the AST as part
of the list of member definitions and will always be the last "member"
of the list, which should help the IDE plugin a lot.