Enhance JFXC-2030
This change improves error recovery for parsing blocks by mkaing sure we
do not place the entire block in to erroneous mode unles we have no other
choice. As per recent changes to class defintinos, the block is created with
what we have, and the final entry in the block_expression will be the error
that caused us to end the block parse. A slight difference between class
definitions arises because the last expression in a block is its return
value.  The tree wlakers have problems if the return element of the block
is Erroneous, soo we prevent this from  happening by setting the return
expression rto null if the last element is Erroneous and placing the
Erroneous node as the last element of the block.