Progress JFXC-2031
When an input sequence consists of extraneous tokens, the parser is
able to auto-recover by reportign the extraneous token but just removing
it from the input stream so that the parsing rules just don't see the
error.

IN certain instances however, the extraneous token is not the one that causes
error detection, but the token before that. This is especially true when
typing in the code using the IDE. Consider:

var
var Answer : 42;

In this case, the parser will see the second 'var' as being extraneous and
therefore report an error against it. However the human sees that the
first variable declaration is not yet complete and really expects the IDE
to know this.

This change then looks for specific cases of extraneous tokens, whereby
the extraneous token is a repeat of teh prior token. If the error routines
detect this condition, they report the error against the first occurrence
of the token, and not the second. As the extraneous token is arbitrarilly
the first or second occurence, this makes no differnce in terms of error
reporting, but 'seems' to make more sense in a lot of cases generated by the
IDE.