Fixed JFXC-2031
Implement scopes in a number of grammar rules such that the error reporting system
knows the context in which to produce erroneous nodes. Simply put, this means
that  Erroneous nodes are always perfectly placed in the AST and are not simply
accumualted at the end of the synthesized function definition. This is exactly
what was asked for in JFXC-2031.

Change recovery in the rules because though the setup is now more complicaed
the actual provision of erroneous nodes is simpler.

Provide much finer grained Erroneous nodes for Object literal and keyframes as
well as eliminate the need for Erroneous blocks and almost always create a
real AST node that conatins Erroneous nodes at the relevant spots. Again, this
should be a lot easier to deal with for the NB plugin.


Please try this and provide feedback, assuming it is good then I will ensure the same
techniques are used on all rules (if I have not in fact already done so).