First version of framework for performance-testing. This is a very rough version, especially the error-reporting is not very user-friendly. I will improve that later, after it drove me crazy enough.
The tool expects a script with the test-cases to run (see below). It runs each tests a number of times and calculates the average of the results. The result is written to a csv-file, which is a comma-separated table and can be imported in OpenOffice etc. If the log-file already exists, the new data is appended. This will give us some nice big tables after some time and allows us to create graphs to document our improvement. The old log-file is copied to <filename>.bak, just in case. To each test-run a description can be attached. We should use that to describe the latest changes, so that we see later which changes caused which improvements (or regressions).

The parameters are
-s <testscript>   Specify test-script
-o <logfile>      Specify log-file
-i <iterations>   Number of iterations
-d <description>  Description
-h                Prints this help


The test-script is a list of commands. Each command has to be in one line and is of this form:
<command> <arg1> <arg2>

Currently the only known command is "SE", which will start the SequenceExerciser. Possible commands are:
SE "10" "stream 10000000"
SE "1000" "stream 2500000"
SE "100000" "stream 75000"

This will run a couple of "use sequence as a queue"-tests with different sizes of the queue. Please note, that for the SeequenceExerciser every argument has to be quoted (as I said, the parsing part is extremely rough at present).