Fixed JFXC-2042 : Centralized null check facility needed by elision General mechanism implemented. Not yet used.
Commented-out partial JFXC-2067 : Use NullChecker for function invocation
NullChecker client for function invocation used in the development of the NullChecker.
Code not quite finished and too great a change for this late in the release.

As the above will address JFXC-2056 : Calling a variable of function type throws NPE if it is null
This represents a partial commented-out fix fo rthis bug as well.
A currently-failing bug has been added for this.