progress, and partially implement, JFXC-2016
Theis submission creates the JFXTree level interface SyntheticTree, which
allows and tree node to indicate to a caller, whether it was generated as
a direct relfeceltion of the source code, or because teh final runnable
class needs the code it will cause to be generated. For instnance, we must
manufacture a Main class node that encapsulates script level variables
and functions etc.


All JFXTree nodes provide this interface, which is essentially:

SynthType getGenType()
setGenType(SynthType)

If a node is not changed from the default via setGenType(), then its
generated type is SynthType.COMPILED, which indicates it was produced
as a direct reflection of compiled code. There is only one other type
at the moment, being SynthType.SYNTHETIC, but other node types may be
added if callers wish to distinguish some more subtle deviation from the
normal COMPILED type.

I have modifed the JFX tree maker such that it flags the two or three
nodes that are obviously synthetic, but require a little more feedback
from the IDE plug in team as to which, (if any) other nodes should be
marked as synthetic in this way. Please try the interface as soon as
possible and append examples of any other nodes that should be flagged as
synthetic to this report, JFXC - 2016. I will not close this until ehte
IDE plugin team think we have correctly set all nodes.

Note that this interface has no effect on code generation or any
other aspect of the tree walkers; it is purely informational and purely
pure the netbeans IDE plugin (or indeed, any other pugin that finds it
useful).