Fixed JFXC-2068 : Use NullChecker for select Introduce SelectTranslator, a subclass of NullChecker, to do all select processing cleanly and with consistent null checks. Fixes numerous problems with select esp for select of a function. With regression test.
Fixed JFXC-2067 : Use NullChecker for function invocation
Use the NullChecker to fix function call problems and clean the code as needed for elision.

These are the two issues in the umbrella (I wasn't able to fix one without fixing the other). So, the umbrella is fixed...
Fixed JFXC-2066 : Standardized NullChecker should be used universally
As a result, the old null checking crud has been deleted.