Fixed JFXC-2065
This fixes the report in JIRA and attempts to predict where other things might
fall over and pre-emptively fix them. The pretty printer was not designed to
print out an erroneous tree, hence it assumes that all nodes are in tact and
that noe of them are erroneous - I have protected the obvious cases from
this eventuallity.