Implement JFXC-2062
Provide parsing support for import of package and all sub-packages.

Note that this change only builds the tree and adds recognition for the
'**' element in JavafxMemberEnter and JavaPretty. Support for this option
within in the tree will be provided by Sundar, who is implementing
the import itself as part of JFXC-1413.