Fixed JFXC-1445
This abort was caused by an error in ANTLR 3.1 (and prior), which I found and
fixed for ANTLR 3.1.1.

This commit changes the buidl to use ANTLR 3.1.1, which has now had a little time to
be checked. You will need to perform ant real-clean and rebuld to be sure of
having nothing left over from ANTLR 3.1.