JFXC-2121 : Need ability to script JavaFX applets from a web page
Added implementations of interfaces exposed in the new Java Plug-In in
Java SE 6 Update 10 to allow JavaScript to call JavaFX Script. Access
to JavaFX Script variables, calling JavaFX Script functions, access to
JavaFX sequences, and both simple and complex kinds of value
conversion are supported. Associated changes are being checked in to
the GUI workspace to hook up this code to the JavaFX applet wrapper;
however, the bulk of the new code is in this workspace. Preliminary
documentation is in doc/liveconnect/README.txt. Not possible to
incorporate test cases in this workspace because of dependence on GUI
libraries; examples will be checked in to the GUI workspace.

Stub jar file containing necessary compile-time dependencies from the
Java Plug-In is being incorporated with this checkin. If the new Java
Plug-In is not available at run-time, the calling code simply catches
any associated NoClassDefFoundErrors and the new functionality is
disabled. The common code paths are untouched and no dependencies
exist from the core JavaFX compiler runtime to this new code.

This represents the last of the supporting work for the recent VP demo.