Fixed JFXC-1766
This change removes fx types that are no longer relevant, at the request of the Netbeans
IDE team, as it is unclear to them what things they need to support when there is old
stuff in there. Because of the proximity of the cut-off I did not remove types and classes
that were widely used in the source code.

This change removes checks for things that are now removed from the code.

This change also finalizes the postion of the lexer/parser for release, in terms of
how it reports on keywords and operators that are no onger supported. In general, warings
are now errors, directing the user at the offical way to do things. The exception is attribute,
which is still a warning because the preview supported attribute and this release would
be a pain if it absolutely forced people to change all their code for that.

Some error messages were elaborated upon, and regression and should-fail tests were updated
to reflect the error message changes.