Introduced new iterator-methods to Sequence: reverseIterator creates an iterator that iterates in reverse order, bound-iterators to iterate only through a part of a sequence.
Created null-safe iterator-methods in Sequences similiar to Sequences.size(). If null is given as argument, an iterator for an empty sequence is returned.

Replaced all loops of the kind "for (i : indexes) sequence.get(i)" with a loop that uses an iterator instead.

This prepares the runtime for JFXC-2150 (implementation of specialized iterators).