Fixed: JFXC-1715
Final check in for this report removes the now redundant classes
JFXInterpolate and InterpolateTree and removes the redundant tag
INTERPOLATE to boot.

Regression tests updated accordingly so that the they do not complain
about the missing INTERPOLATE type.

Note: External developers will need to remove code referring to this tag
too.