Fixed - JFXC-2172
The paresr generated correct nodes for import x.** but the support
for this was remvoed from the runtime without protecting it against
the parser generating it. The parser now therefore replaces ** with * so that
code completion and so on does not not cause aborts in the AST code, but
issues an error message to show that the ** part is currently not valid
syntax.