Partially implement JFXC-2187
This change implements the first part of allowing sequences of functions as well as funcitons returning sequences.

The change allows parenthesis to be used to associate or dissassociate cardinality operators
when specifying types. So, before this change you coudl have only:

var x : function(): String [];

 - x is a var of type 'function returning a sequence of strings'

Whereas now you can additionally have:

var y : (function() : String)[] ;

 - y is var of type 'sequence of functions returning String'

However, while this change allows the syntax of the form, it does not implement the
necessary changes to support such a declaration. Some testing of various type specifications
shows that this change does not break any existing types and the newly specifiable types
seem to result in reasonable error mesesages from the translation to java phase. Hence it is
checked in now, so that we can determine the correct tree shapes.