Fixing RT-1145 - Doc tool does not support package level documentation.
Minimal package-level doc support is added. In the overview page,
currently we show a one-liner doc for each package. When there is
no package-info.java or package-info.fx, the overview page showed
one-liner for first class in the package. This is fixed now. When no
package doc, nothing is shown now.

Also, when there is detailed package level doc via package-info.fx/.java,
the overview page shows an arrow button. On clicking the arrow button,
the detailed doc is shown "in place" - much like method doc expansion for
in class doc page.