Fixed JFXC-2186
The error checking routine that evaluates whether a semi colon
is required to separate a statement was not protected against
being positioned at the first token in the input stream when the input
text was only a single token such as 'function'. This change now
protects these routines against this event.