Partial fix for JFXC-2443 - Get rid of all the incorrect assumptions in attribution about the numeric types.
It is still "partial" because conversions of sequences are yet to be
handled.

This fix reverses changes done by r4720 for JFXC-2445 : Float arrays not
properly type-inferenced, breaks GUI issue. Added relevant methods in
Sequences class (to convert float[] to Sequence<Float> etc.) so that the
previous crash during translation is now avoided.