Fixed JFXC-2608
String expressions always insert an imaginary string format node
into the tree for STRING_EXPRESSION. However, the node was being
marked up as if it were a single character string as far as the IDE
was concerned, when in fact it was a zero length string in terms of
the extant source code that the IDE deals with. This change therefore
marks the string formatting node as being an empty string source code
wise. The code genrator knows to ignore this empty string already and we
need it so that we can easily process multi expression STRING_EXPRESSIONs
by being able to rely on a consistent number of elements in the expression
collection that the code generator will eventually process.

After this change, the IDE can process STRING_EXPRESSION trees in the same
consistent manner as the code generator and can detect that the format
string is imaginary by noticing that the start and end positions are
the same.