Fixed: JFXC-2294
When the name of a function was left out of a feunction declaration
we will get 2 errors. One says that the identifier is missing and the
second says that the function should have a valid name. However, the
second error message was being reported in an incorrect position and
so was confusing. We now get one error message saying the identifier
is missing and a second, more specific message saying that declared
functions must be named (against the correct part of the source code).