Improved: JFXC-2677
Prior to this change, missing ',' delimiters in sequences and similar would
give a message that was OK, but not great, about how it was looking for a
sequene expression but found 'identifier' or something similar. This
modification improves the error message to specifically state that a comma
delimiter is required or that perhaps the programmer has a mal-formed expression
which should siginificantly reduce confusion about what the compiler is
expecting.