Fix JFXC-3056 : remove intermediate Locations from bind translation - infrastructure Implement a translation mechanism that allow optionality in the intermediate results of bound expression translation. Change the fundamental assumption in bound expression transformation that each visitor and each translation method will yield a JCExpression whose runtime value is a Location, to instead yield an adapter object that can optional give a JCExpression(Location), a JCExpression(BindingExpression), or a BindingExpressionClosureTranslator (which could be folded into outer translations).
Fix JFXC-3057 : remove intermediate Locations from bind translation - top-level
The result of most bound translations are bound to a variable. Currently, the Location-only bound translation approach means that many bindings have an extra layer of XxxVariable and binding expression at the point they are applied to the bind.
For example: def bb = bind Foo{};
is translated as:
   $bb.bind(false, ObjectVariable.<u.Foo>make(null, false, new _SBECL(0, null, null, null, 0)));
the binding expression that codes for the instanciation of Foo is wrapped in an ObjectVariable to make it a Location.
Internally, this will be wrapped in another binding expression, since that is what binding needs.
After this optimization, the translation would be:
   $bb.bind(false, new _SBECL(0, null, null, null, 0));
The extra Variable and binding expression are gone, improving execution performance, and dynamaic and static footprint.