Fixed JFXC-3110
The ANTLR tool has a timeout parameter that causes the conversion of NFA
to DFA to timeout after a period of time. As teh JVM does not allow capture
of CPU time used by a thread, this timeout is absolute, which has the unfortunate
side effect that on machines that are very busy, or are much slower than the
machine where the grammar was developed, that the build can timeout and fail.

Hence this patch increases the absolute timeout of the analysis for v4parser.g
to be 10 seconds rather than 1 second, which should prevent the timeouts observed
in some of the Hudson builds.