Partially Fixed: JFXC-2963
Some error messages could end up needlessly repeating the name/type of element
that was in error, the most annoying of which was:

Sorry, I was trying to understand an identifer when I found an extra identifier which is an identifier that should not be there.

and similar. This change does not address all the potential changes mentioned in JFXC-2963
as those must neccesarilly be discussed in wider circles.