JFXC-3202 use wildcard more consistently for sequence value types
Fixes some inconsistency in how the sequence type T[] is translated
to Sequence<? extends T>, specifically in te context of the current
value of a SequenceVariable.

This commit is rushed - it compiles and build the runtime, but some
of the internal tests (written in Java) fail.  They will just
have to fixed en masse.