Added facility to compile .javadump files generated with -XDdumpjava option. When javafxc is invoked with the (hidden) compiler option -XDmjavac, the compiler accepts .javadump files (generated by earlier -XDdumpjava invocation and possibly manually edited afterwards) and compiles those files.
To faciliate this, JavaPretty is changed slightly (a) to add import
declarations for runtime packages and (b) to add a semicolon after value
expressions of block expressions.

Reviewed by: Maurizio

Verification:

* Wrote simple .fx classes and compiled with -XDdumpjava to generate javadumps.

* Compiled the .javadump files generated in above step with -XDmjavac
invocation.

* Also, checked that block expressions are properly compiled. For now,
manually edited .javadump file to add a block expression in a getter method.
Checked the resultant .class files with javap and by running the .class file.