Progress JFXC-3515 :   Compiled bind: bound functions.
For local vars with bind/on-replace, JavafxLocalToClass already
generates a class. For bound functions with one or more vars (which
are implicitly bind). Introducing a synthetic var for result expression
so that a class is generated for bound functions always. Checked that
a class is generated by temporarily turning on local-to-class and invoking
compiler with -XDdumpinflate option. (turned off local-to-class before this
commit).