Progress JFXC-3561 : Compiled bind: Tests that do not compile.
Fixed bug 6: Extending a class and overriding a script level func fails
   test/regress/jfxc1245B.fx
    [junit] test\regress\jfxc1245B.fx:11: access$scriptLevel$() in jfxc1245B cannot override access$scriptLevel$() in jfxc1245Bsub; attempting to use incompatible return type
    [junit] found : jfxc1245B.jfxc1245B$Script
    [junit] required: jfxc1245Bsub.jfxc1245Bsub$Script
    [junit] class jfxc1245B extends jfxc1245Bsub {}
    [junit] ^

The compiler generated access$scriptLevel$ method with same name but with
different return type for a class and it's subclass -- which is rejected by
the back-end. The fix is to concatenate the fully qualified class name with
the prefix "access$scriptLevela$".

Testing: Ran "ant dev-pass" on Mac OS X. Ran "ant dev-fail dev-check" to
adjust project.properties.