The test jfxc1404.fx failed to compile because of type mismatch (Object assigned to String without cast). This test passes "null" to a String type parameter of a Java method call inside a bind expression.
In decompose, we shred the var with the type of initialization expression.
If "null" is passed to a function call, we shred the arg expression for "null"
as a variable like:

    var tmp = null; // with <nulltype>

The above variable's type becomes java.lang.Object in the translator. Which
can not be assigned to String. The fix is to use formal parameter types of
the called function for such cases.

With this change, jfxc1404.fx compiles, but now we get a NPE thrown from the
Java method called from a bind expression. In the location world, we had a
central place in runtime code where we wrapped "compute" call with a try..catch
to call ErrorHandler.handleException(e). This was bind exceptions are silently
eaten away. In the compiled-bind world, we need to generate try..catch in every
getter method body.