Refine JFXC-3627 : Compiled bind: bound sequence: function invocation Handle on-replace correctly for bound invocation of a function returning a sequence.
Fixed JFXC-3676 : Compiled bind: applyDefaults and VarInit
The interaction between VarInit and initialization has been adhoc, and was getting worse.
I ran head long into this in implementing the above -- it worked at the instance level, but not the script level.
So, I've finally made the behavior consistent. ALL classes run initialize$ (unless this is done manully -- object literals).
applyDefaults$ now by-passes initialization of vars with a VarInit. It is re-enabled by VarInit execution. This is a new flag.
Side-efect is that forward-reference is handled consistently: forward-reference in a linear execution context (block (even if inflated),
or script) does not initialize the forward value (this corresponds to the case where there is a warning); whereas, forward-reference in
a class instance (even subclassed and overridden) initializes an uninitialized var.

Added regression tests --
 test\regress\jfxc3627b.fx
 test\regress\jfxc3627onra.fx
 test\regress\jfxc3627onrb.fx
 test\regress\jfxc3627onrc.fx
 test\regress\jfxc3627onrd.fx
 test\regress\jfxc3676a.fx
 test\regress\jfxc3676b.fx
 test\regress\jfxc3676c.fx
 test\regress\jfxc3676forward.fx

Was pass, now runtime failure (I've filed JFXC-3681) --
 test/features/F20-bound-functions/SeqBoundFunc.fx \