Progress 3555 : Compiled bind: Tests that compile ok but fail at runtime.
Fixed bug 15 of JFXC-3555. The test jfxc2788and.fx has a function "here"
that calls "println". The function is called inside a bind expression.
The expected file for this test expects output lines from that println.

Since we assume that the functions called from bind expr. are side-effect free,
we use the cached value when arguments and the receiver don't change. Removed
the println in the function and adjusted the .EXPECTED file.

Testing : ran "ant dev-pass" on Mac OS X 10.5.8.